tools/PokeemeraldMapConverter/requirements.txt
# Python dependencies for pokeemerald to Tiled conversion tools
# Install with: pip install -r requirements.txt

Pillow>=10.0.0
